+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,13 @@
 # druxt-blocks
 
+## 0.14.1
+
+### Patch Changes
+
+- Updated dependencies [9161b38]
+  - druxt-router@0.24.0
+  - druxt-entity@0.21.2
+
 ## 0.14.0
 
 ### Minor Changes