+++ packages/site/CHANGELOG.md
@@ -1,5 +1,16 @@
 # druxt-site
 
+## 0.10.1
+
+### Patch Changes
+
+- Updated dependencies [9161b38]
+  - druxt-router@0.24.0
+  - druxt-blocks@0.14.1
+  - druxt-breadcrumb@0.13.1
+  - druxt-entity@0.21.2
+  - druxt-views@0.16.1
+
 ## 0.10.0
 
 ### Minor Changes